split utils/test_files into utils/r and utils/files
Extracts rStringLiteral / rPathLiteral into utils/r.ts since they encode
TypeScript values for R interpolation and aren't file-specific. Renames
test_files.ts to files.ts because writeAndOpenFile / closeAndDeleteSandboxFiles
aren't test-only -- the 'test' in the name was scope, not domain.
